requirements.txt
ccxt
requests
python-dotenv
cryptography